social_media_trading_platform/trading_platform/requirements.txt
django
djangorestframework
psycopg2
djangorestframework-simplejwt
django-cors-headers
channels
redis
requests
